refactor: update theme type imports to use new path and remove deprecated theme-types file
