+++ src/theme/presets/amethyst-haze.ts
@@ -1,4 +1,4 @@
-import type { ThemePreset } from "../theme-types";
+import type { ThemePreset } from "@/types/theme-types";
 
 export const amethystHaze: ThemePreset = {
   label: "Amethyst Haze",
