+++ src/theme/presets/bold-tech.ts
@@ -1,4 +1,4 @@
-import type { ThemePreset } from "../theme-types";
+import type { ThemePreset } from "@/types/theme-types";
 
 export const boldTech: ThemePreset = {
   label: "Bold Tech",
