+++ src/theme/presets/clean-slate.ts
@@ -1,4 +1,4 @@
-import type { ThemePreset } from "../theme-types";
+import type { ThemePreset } from "@/types/theme-types";
 
 export const cleanSlate: ThemePreset = {
   label: "Clean Slate",
