+++ src/theme/presets/cosmic-night.ts
@@ -1,4 +1,4 @@
-import type { ThemePreset } from "../theme-types";
+import type { ThemePreset } from "@/types/theme-types";
 
 export const cosmicNight: ThemePreset = {
   label: "Cosmic Night",
