+++ src/theme/presets/default-neutral.ts
@@ -1,4 +1,4 @@
-import type { ThemePreset } from "../theme-types";
+import type { ThemePreset } from "@/types/theme-types";
 
 export const defaultNeutral: ThemePreset = {
   label: "Default Neutral",
