+++ src/theme/presets/default-red.ts
@@ -1,4 +1,4 @@
-import type { ThemePreset } from "../theme-types";
+import type { ThemePreset } from "@/types/theme-types";
 
 export const defaultRed: ThemePreset = {
   label: "Default Red",
