+++ src/theme/presets/default-rose.ts
@@ -1,4 +1,4 @@
-import type { ThemePreset } from "../theme-types";
+import type { ThemePreset } from "@/types/theme-types";
 
 export const defaultRose: ThemePreset = {
   label: "Default Rose",
