+++ src/theme/presets/default-stone.ts
@@ -1,4 +1,4 @@
-import type { ThemePreset } from "../theme-types";
+import type { ThemePreset } from "@/types/theme-types";
 
 export const defaultStone: ThemePreset = {
   label: "Default Stone",
