+++ src/theme/presets/default-violet.ts
@@ -1,4 +1,4 @@
-import type { ThemePreset } from "../theme-types";
+import type { ThemePreset } from "@/types/theme-types";
 
 export const defaultViolet: ThemePreset = {
   label: "Default Violet",
