+++ src/theme/presets/default-zinc.ts
@@ -1,4 +1,4 @@
-import type { ThemePreset } from "../theme-types";
+import type { ThemePreset } from "@/types/theme-types";
 
 export const defaultZinc: ThemePreset = {
   label: "Default Zinc",
