+++ src/theme/presets/doom-64.ts
@@ -1,4 +1,4 @@
-import type { ThemePreset } from "../theme-types";
+import type { ThemePreset } from "@/types/theme-types";
 
 export const doom64: ThemePreset = {
   label: "Doom 64",
