+++ src/theme/presets/midnight-bloom.ts
@@ -1,4 +1,4 @@
-import type { ThemePreset } from "../theme-types";
+import type { ThemePreset } from "@/types/theme-types";
 
 export const midnightBloom: ThemePreset = {
   label: "Midnight Bloom",
