+++ src/theme/presets/mocha-mousse.ts
@@ -1,4 +1,4 @@
-import type { ThemePreset } from "../theme-types";
+import type { ThemePreset } from "@/types/theme-types";
 
 export const mochaMousse: ThemePreset = {
   label: "Mocha Mousse",
