+++ src/theme/presets/modern-minimal.ts
@@ -1,4 +1,4 @@
-import type { ThemePreset } from "../theme-types";
+import type { ThemePreset } from "@/types/theme-types";
 
 export const modernMinimal: ThemePreset = {
   label: "Modern Minimal",
