+++ src/theme/presets/northern-lights.ts
@@ -1,4 +1,4 @@
-import type { ThemePreset } from "../theme-types";
+import type { ThemePreset } from "@/types/theme-types";
 
 export const northernLights: ThemePreset = {
   label: "Northern Lights",
