+++ src/theme/presets/ocean-breeze.ts
@@ -1,4 +1,4 @@
-import type { ThemePreset } from "../theme-types";
+import type { ThemePreset } from "@/types/theme-types";
 
 export const oceanBreeze: ThemePreset = {
   label: "Ocean Breeze",
