+++ src/theme/presets/quantum-rose.ts
@@ -1,4 +1,4 @@
-import type { ThemePreset } from "../theme-types";
+import type { ThemePreset } from "@/types/theme-types";
 
 export const quantumRose: ThemePreset = {
   label: "Quantum Rose",
