+++ src/theme/presets/retro-arcade.ts
@@ -1,4 +1,4 @@
-import type { ThemePreset } from "../theme-types";
+import type { ThemePreset } from "@/types/theme-types";
 
 export const retroArcade: ThemePreset = {
   label: "Retro Arcade",
