+++ src/theme/presets/rose-gold.ts
@@ -1,4 +1,4 @@
-import type { ThemePreset } from "../theme-types";
+import type { ThemePreset } from "@/types/theme-types";
 
 export const roseGold: ThemePreset = {
   label: "Rose Gold",
