+++ src/theme/presets/solar-dusk.ts
@@ -1,4 +1,4 @@
-import type { ThemePreset } from "../theme-types";
+import type { ThemePreset } from "@/types/theme-types";
 
 export const solarDusk: ThemePreset = {
   label: "Solar Dusk",
